+++ src/index.js
@@ -15,7 +15,7 @@ export {
   BASE_FEE
 } from './transaction_builder';
 export { Asset } from './asset';
-export { ChangeTrustAsset } from './change_trust_asset';
+export { LiquidityPoolAsset } from './liquidity_pool_asset';
 export { TrustLineAsset } from './trustline_asset';
 export {
   Operation,